+++ src/reuse/_annotate.py
@@ -54,27 +54,27 @@ _LOGGER = logging.getLogger(__name__)
 
 
 def verify_paths_line_handling(
+    args: Namespace,
     paths: Iterable[Path],
-    parser: ArgumentParser,
-    force_single: bool,
-    force_multi: bool,
 ) -> None:
-    """This function aborts the parser when *force_single* or *force_multi* is
+    """This function aborts the parser when --single-line or --multi-line is
     used, but the file type does not support that type of comment style.
     """
     for path in paths:
-        style = _get_comment_style(path)
+        style = NAME_STYLE_MAP.get(args.style)
+        if style is None:
+            style = _get_comment_style(path)
         if style is None:
             continue
-        if force_single and not style.can_handle_single():
-            parser.error(
+        if args.single_line and not style.can_handle_single():
+            args.parser.error(
                 _(
                     "'{path}' does not support single-line comments, please"
                     " do not use --single-line"
                 ).format(path=path)
             )
-        if force_multi and not style.can_handle_multi():
-            parser.error(
+        if args.multi_line and not style.can_handle_multi():
+            args.parser.error(
                 _(
                     "'{path}' does not support multi-line comments, please"
                     " do not use --multi-line"
@@ -123,9 +123,10 @@ def add_header_to_file(
     """Helper function."""
     # pylint: disable=too-many-arguments,too-many-locals
     result = 0
-    if style is not None:
-        comment_style: Optional[Type[CommentStyle]] = NAME_STYLE_MAP.get(style)
-    else:
+    comment_style: Optional[Type[CommentStyle]] = NAME_STYLE_MAP.get(
+        cast(str, style)
+    )
+    if comment_style is None:
         comment_style = _get_comment_style(path)
     if comment_style is None:
         if skip_unrecognised:
@@ -487,13 +488,7 @@ def run(args: Namespace, project: Project, out: IO[str] = sys.stdout) -> int:
         verify_write_access(paths, args.parser)
 
     # Verify line handling and comment styles before proceeding
-    if args.style is None and not args.force_dot_license:
-        verify_paths_line_handling(
-            paths,
-            args.parser,
-            force_single=args.single_line,
-            force_multi=args.multi_line,
-        )
+    verify_paths_line_handling(args, paths)
 
     template, commented = get_template(args, project)