+++ CHANGELOG.md
@@ -81,6 +81,8 @@ CLI command and its behaviour. There are no guarantees of stability for the
 ### Fixed
 
 - Reduced python-debian minimum version to 0.1.34. (#808)
+- Fix issue in `annotate` where `--single-line` and `--multi-line` would not
+  correctly raise an error with an incompatible comment style. (#853)
 
 ### Security